% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ms}
\alias{ms}
\title{Create a multistate event object}
\usage{
ms(id, time, status)
}
\arguments{
\item{id}{A vector of id variable.}

\item{time}{A vector of follow-up times.}

\item{status}{A vector of event type, \code{k} if transitioning to state \eqn{k},
0 if censored and \eqn{K+1} represents death.}
}
\value{
An object of class \code{ms} used as an argument for \code{\link{rmtfit}}.
}
\description{
Create a multistate event object
}
