% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.rmtfit}
\alias{print.rmtfit}
\title{Print a short summary of rmtfit objects}
\usage{
\method{print}{rmtfit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{rmtfit}}.}

\item{...}{Further arguments passed to or from other methods}
}
\value{No return value, called for side effects.}
\description{
Print the results for the restricted mean times in favor of treatment.
}
