% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.summary.rmtfit}
\alias{print.summary.rmtfit}
\title{Print method for summary.rmtfit objects}
\usage{
\method{print}{summary.rmtfit}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{summary.rmtfit}}.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
No return value, called for side effects.
}
\description{
Produces a printed summary of the results for the restricted
mean times in favor of treatment
}
