% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rec}
\alias{rec}
\title{Create a recurrent event object}
\usage{
rec(id, time, status)
}
\arguments{
\item{id}{A vector of id variable.}

\item{time}{A vector of follow-up times.}

\item{status}{A vector of event type, 1 = recurrent event, 2 = death, and 0 = censoring;}
}
\value{
An object of class \code{rec} used as an argument for \code{\link{rmtfit}}.
}
\description{
Create a recurrent event object
}
