\name{RMT-package}
\alias{RMT-package}
\alias{RMT}
\docType{package}
\title{
read counts and rpkm for MeRIP-seq data
}
\description{
The package is developed for the analysis of RNA epigenome sequencing data from MeRIP-seq or m6A-se.

The RNA modification tools R-package accepts and statistically supports multiple biological replicates, internally removes PCR artifacts and multiple mapping reads, outputs exome-based binding sites and reads counts and rpkm in every bams. Besides, the package compute the fold change in ip and input bams and select features.The package is still under active development, and we welcome all biology and computation scientist for all kinds of collaborations and communications. 

Please feel free to contact Dr. Lian Liu <liulian19860905@163.com>if you have any questions.
}
\details{
\tabular{ll}{
Package: \tab RMT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-03\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Lian Liu, PhD 
Maintainer: <liulian19860905@163.com>
}

\keyword{ RNA epigenetics }

\examples{

library(RMT)
?RMT
}

