% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmt-package.R
\docType{package}
\name{rmt-package}
\alias{rmt}
\alias{rmt-package}
\title{rmt: Restricted Mean Time in Favor of Treatment}
\description{
Contains inferential and graphical routines for comparing two treatment arms in terms of the restricted mean time in favor of treatment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/view/lmaowisc/}
  \item \url{https://lmaowisc.github.io/rmt/}
}

}
\keyword{internal}
