\name{SN.diagram}
\alias{SN.diagram}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Stress-Life (S-N) Diagram
}
\description{
Generates the stress-life (or S-N) diagram based on various stress and fatigue life input data.
}
\usage{
SN.diagram(input_type,dat,stressunits,options)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input_type}{
Numeric definition for the input type of the data and its options: 1 - Fully reversed (R = -1) S-N data points, 2 - cyclic stress range , 3 - multi-axial stress
}
\item{data}{
Data is entered as a list of fully reversed stress and cycle data.  List may include runoff stress and cycle data. (Order of list: (1) alternating stress, (2) cycles to failure, (3) runoff stress, (4) runoff cycles).  If no data is to be entered and stress-life parameters \eqn{A} and \eqn{b} are entered in \code{options}, then \code{NA} should be entered here.
}
\item{stressunits}{
Defines the units for the stress amplitude: \code{1} for \emph{'MPa'} and \code{2} for \emph{'ksi'}.  \emph{'MPa'} will be used as default if this heading is left empty.
}
\item{options}{
Optional input vector made up of information known about the material primarily.  This heading may be left blank if data is provided, in which case an estimate of \eqn{Su} and endurance limit \eqn{Se} based on the given data.  Options include:
  \itemize{
  \item \strong{A} - Life-stress parameter of S-N equation \eqn{S = A (N_f^b)}
  \item \strong{b} - Life-stress parameter of S-N equation \eqn{S = A (N_f^b)}
  \item \strong{Su} - Ultimate strength
  \item \strong{Se} - Endurance limit
  \item \strong{BHN} - The Brinnell Hardness Number
  \item \strong{corr_rel} - The correlation relationship between mean stress \eqn{S_m} and stress amplitude \eqn{S_a} and equivalent stress amplitude \eqn{S{ar}} for non-zero mean stress (or non-fully reversed loading conditions).
  \enumerate{
    \item \strong{Soderberg relationship} - (enter as "Soderberg")
    \deqn{\frac{S_a}{S_{ar}} + \frac{S_m}{S_y} = 1} (\emph{NOTE}: \eqn{S_y} yield stress to be entered in list of \code{options})
    \item \strong{Modified Goodman relationship} - (enter as "ModifiedGoodman")
    \deqn{\frac{S_a}{S_{ar}} + \frac{S_m}{S_u} = 1}
    \item \strong{Morrow relationship} - (enter as "Morrow")
    \deqn{\frac{S_a}{S_{ar}} + \frac{S_m}{\sigma_f} = 1} (\emph{NOTE}: \eqn{\sigma_f} (\code{sigf}) true fracture strength to be entered in list of \code{options})
    \item \strong{Gerber relationship} - (enter as "Gerber")
    \deqn{\frac{S_a}{S_{ar}} + (\frac{S_m}{S_u})^2 = 1}
    \item \strong{Smith-Watson-Topper relationship} - (enter as "SWT")
    \deqn{\frac{S_a}{S_{ar}} = (\frac{1 - R}{2})^{0.5}}
    \item \strong{Walker relationship} - (enter as "Walker")
    \deqn{\frac{S_a}{S_{ar}} = (\frac{1 - R}{2})^{1 - \gamma}} (\emph{NOTE}: \eqn{\gamma} (\code{gam}) the fitting constant to be entered in list of \code{options})
  }
  \item \strong{Srange} - Axial cyclical loading stress range vector made up of a maximum and minimum stress under non-fully reversed conditions (ex. \code{c(max stress,min stress)}).
  \item \strong{Srange (normal stress multi-axial)} - In the case of multi-axial loading stress, enter the multi-axial forms of \strong{Srange}: \emph{Srangex}, \emph{Srangey}, and \emph{Srangez}.  Alternatively, if the multi-axial normal amplitudes and mean stresses are already known, enter instead: \emph{sig_xa}, \emph{sig_ya}, \emph{sig_za}, \emph{sig_xm}, \emph{sig_ym}, and/or \emph{sig_zm}.
  \item \strong{Trange (shear stress multi-axial)} - Multi-axial loading stress may include shear stress activity along planar regions x-y, y-z, and x-z. Enter these multi-axial ranges if known as: \emph{Trangexy}, \emph{Trangeyz}, and \emph{Trangexz}.  Alternatively, if the multi-axial shear amplitudes stresses are already known, enter instead: \emph{tau_xya}, \emph{tau_yza}, and/or \emph{tau_xza}.
  \item \strong{Strace} - Desired stress value that user requires the matching life from based on the S-N curve.
  \item \strong{Ntrace} - Desired life cycle value that user requires the matching alternating stress from based on the S-N curve.
  }
}
\item{confid}{
The confidence in the S-N curve estimates entered as a probability between \eqn{0} and \eqn{1}.  Default is \eqn{0.95} (or \eqn{95\%})
}
}
\details{
The S-N diagram generated by this code is based primarily on the available data and/or the optional material and test properties in addition to the computational bases by which the S-N diagram properties are estimated.  The log-log line depicting failure by stress for example is estimated by either available stress-life data or the known life-stress parameters \eqn{A} and \eqn{b} (entered as \code{options}).
}
\value{
The code will produce the S-N diagram with a 95\% confidence bound about the mean S-N curve (\emph{if} more than two data is provided).  The command window will print the estimate for the S-N curve equation (valid between \eqn{10^3} and \eqn{10^6} fatiigue cycles) and the estimated endurance limit \eqn{Se}.  If a trace for stress or life is requested, the life and stress estimates for those respective traces will print as well.  A list will also be output in the following order:
\item{SNdiag}{
  The S-N diagram.
}
\item{A}{
  The estimate (or given value) for stress-life parameter \eqn{A}.
}
\item{b}{
  The estimate (or given value) for stress-life parameter \eqn{b}.
}
\item{Se}{
  The estimate (or given value) for endurance limit \eqn{Se}.
}
\item{Su}{
  The estimate (or given value) for the ultimate strength \eqn{Su}.
}
\item{Strace or Sar}{
  The given value for the stress for which a life estimate is required.  When \code{input_type} is \code{2} or \code{3}, this value defaults to the equivalent stress amplitude \eqn{Sar} for non-zero mean stress.
}
\item{Nequiv}{
  The estimated value of life in cycles at \code{Strace} or \eqn{S_{ar}} \code{Sar}.
}
\item{Ntrace}{
  The given value for the fatigue life cycle for which an alternating stress estimate is required.
}
\item{Sequiv}{
  The estimated value of alternating stress at \code{Ntrace}.
}
}
\references{
Modarres, M., Amiri, M., Jackson, C.  (2017)  Probabilistic Physics of Failure Approach to Reliability: Modeling, Accelerated Testing, Prognosis and Reliability Assessment.  Center for Risk and Reliability

Banantine, J., Corner, J., & Handrock, J. (1990). Fundamentals of Metal Fatigue Analysis. Prentice Hall
}
\author{
Reuel Smith (2022)
}
\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
# Example 1: Consider a set of stress/life steel test data, where these test data are all fully-reversed:
# Failure Stress (MPa): 340,300,290,275,260,255,250,235,230,220,215,210
# Failure Cycles: 15e3,24e3,36e3,80e3,177e3,162e3,301e3,290e3,361e3,881e3,1.3e6,2.5e6
# Runoff Stress (MPa): 210,210,205,205,205
# Runoff Cycles: 1e7,1e7,1e7,1e7,1e7
# List amplitude stress (in MPa) and cycles to failure.
datSN1 <- list(c(340,300,290,275,260,255,250,235,230,220,215,210),c(15e3,24e3,36e3,80e3,177e3,162e3,301e3,290e3,361e3,881e3,1.3e6,2.5e6),c(210,210,205,205,205),c(1e7,1e7,1e7,1e7,1e7))
SN.diagram(1,datSN1,1)

# Example 2: (Problem 1.2a in Bannantine et. al.) Consider a set of stress/life steel test data, where these test data are all fully-reversed:
# Failure Stress (ksi): 32.3,30.3,27.9,25.9,25.4,24.4
# Failure Cycles: 4.5e4,2.4e5,8e5,1.5e6,2.7e6,7.8e6
# Runoff Stress (ksi): 24.4,24,23.6,23.5
# Runoff Cycles: 1e7,2.6e7,1.2e7,2.2e7
# List amplitude stress (in ksi) and cycles to failure.  Also includes known ultimate strength of 42 ksi and BHN of 69.
datSN_MatA <- list(c(32.3,30.3,27.9,25.9,25.4,24.4),c(4.5e4,2.4e5,8e5,1.5e6,2.7e6,7.8e6),c(24.4,24,23.6,23.5),c(1e7,2.6e7,1.2e7,2.2e7))
Su_MatA <- 42
BHN_MatA <- 69
SN.diagram(1,datSN_MatA,2,list(Su = Su_MatA, BHN = BHN_MatA))

# Example 3: (Problem 1.2c in Bannantine et. al.) Consider a set of stress/life steel test data, where these test data are all fully-reversed:
# Failure Stress (ksi): 110,105,100,98,95
# Failure Cycles: 2.4e4,3.1e4,4.5e4,8.7e4,1.5e5
# Runoff Stress (ksi): 92,90
# Runoff Cycles: 1e8,1e8
# List amplitude stress (in ksi) and cycles to failure.  Also includes known ultimate strength of 180 ksi and BHN of 370.  Find the stress amplitude at 10^5 cycles and the cycles to failure when stress amplitude is 115 ksi.
datSN_MatB <- list(c(110,105,100,98,95),c(2.4e4,3.1e4,4.5e4,8.7e4,1.5e5),c(92,90),c(1e8,1e8))
Su_MatB <- 180
BHN_MatB <- 370
SN.diagram(1,datSN_MatB,2,list(Su = Su_MatB, BHN = BHN_MatB, Ntrace = 10^5, Strace = 115))

# Example 4: (Example 2.1 in "Probabilistic Physics of Failure Approach to Reliability: Modeling, Accelerated Testing, Prognosis and Reliability Assessment.")
# Consider a steel component that is subjected to a cyclic load where maximum stress is 760 MPa and minimum stress is 70 MPa.  The steel component has an ultimate stress of 1030 MPa and an endurance limit of 410 MPa.  A fully reversed stress of 760 MPa at 1000 cycles is applied.  If we assume that the Modified Goodman relationship is used for mean stress correction, find the S-N diagram.
SN.diagram(2,list(760,1000),1,list(Su = 1030, Se = 410, Srange = c(760,70), corr_rel = "ModifiedGoodman"))

# Example 5: (Example 2.2 in "Probabilistic Physics of Failure Approach to Reliability: Modeling, Accelerated Testing, Prognosis and Reliability Assessment" an application of multi-axial stress loading.)
# "A pressure vessel with a 1.4 m diameter and 3.6 m length has a wall thickness of 16 mm made of  AISI 1040 HR steel (Sy=227 MPa and Su=415 MPa).  The vessel is subjected to fluctuation of 30 cycles per minute.  In each cycle, the pressure fluctuates between 1 and 5 MPa.  The life of the vessel (i.e number of cycles to rupture) needs to be estimated (assuming the vessel is thin-walled).  The values of A and B for life-stress formulation are 886 and -0.14, respectively."
SN.diagram(3,list(886*(c(1000,10^6)^-0.14),c(1000,10^6),886*(c(10^6,10^6)^-0.14),c(10^7,10^7)),1,list(Su = 415, Sy = 227, sig_xa = 43.8, sig_ya = 87.5, sig_xm = 65.6, sig_ym = 131.3, corr_rel = "ModifiedGoodman"))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
